fix(src): declaration
styled-components need require `react` when generate `.d.ts` files
 move `style.ts` into `*.tsx`
