+++ package.json
@@ -27,13 +27,13 @@
   },
   "dependencies": {
     "react-modal": "^3.4.4",
-    "react-popper": "^1.0.0-beta.6",
+    "react-popper": "^1.0.0",
     "react-slick": "^0.23.1"
   },
   "devDependencies": {
     "@types/jest": "^22.2.2",
-    "@types/react": "^16.3.5",
-    "awesome-typescript-loader": "^5.0.0",
+    "@types/react": "^16.4.0",
+    "awesome-typescript-loader": "^5.1.1",
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.4",
     "enzyme": "^3.3.0",
@@ -43,19 +43,19 @@
     "jest-styled-components": "^5.0.1",
     "lint-staged": "^4.0.3",
     "npm-run-all": "^4.1.2",
-    "react": "^16.3.1",
-    "react-dom": "^16.3.1",
-    "react-hot-loader": "^4.0.1",
-    "styled-components": "^3.2.6",
-    "stylelint": "^9.2.0",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
+    "react-hot-loader": "^4.3.3",
+    "styled-components": "^3.3.2",
+    "stylelint": "^9.3.0",
     "stylelint-config-recommended": "^2.1.0",
     "stylelint-config-styled-components": "^0.1.1",
     "stylelint-processor-styled-components": "^1.3.1",
     "ts-jest": "^22.4.2",
     "tslint": "^5.9.1",
-    "tslint-config-zcool": "^1.6.0",
-    "typescript": "^2.8.1",
-    "webpack": "^4.5.0",
+    "tslint-config-zcool": "^1.7.0",
+    "typescript": "^2.9.2",
+    "webpack": "^4.12.0",
     "webpack-dev-server": "^3.1.3"
   },
   "scripts": {
