--- src/collapse/index.d.ts
@@ -1,2 +1,0 @@
-export { default as Collapse } from './collapse'
-export { default as CollapseItem } from './collapseItem'
