+++ src/globals/index.ts
@@ -1,3 +1,4 @@
+import * as React from 'react'
 import styled, { css } from '../theme/styled-components'
 import { T } from '../util'
