+++ src/list/index.tsx
@@ -1,3 +1,4 @@
+import * as React from 'react'
 import styled from '../theme/styled-components'
 import { T } from '../util'
