+++ src/theme/styled-components.ts
@@ -1,3 +1,4 @@
+import * as React from 'react'
 import * as styledComponents from 'styled-components'
 import { ThemedStyledComponentsModule } from 'styled-components'
