+++ src/tooltip/index.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react'
 import Portal from '../portal'
 import { Manager, Popper, Reference } from 'react-popper'
-import { ToolTipContainer, ReferenceContainer, PopperContainer } from './style'
+import styled, { keyframes } from '../theme/styled-components'
+import { T } from '../util'
 
 interface Props {
   placement:
@@ -45,15 +46,9 @@ class Tooltip extends React.Component<Props, State> {
     const { placement, children, title, className } = this.props
 
     return (
-      <ToolTipContainer onMouseEnter={this.handleEnter} onMouseLeave={this.handleLeave} >
+      <ToolTipContainer onMouseEnter={this.handleEnter} onMouseLeave={this.handleLeave}>
         <Manager>
-          <Reference>
-            {({ ref }) => (
-              <ReferenceContainer innerRef={ref}>
-                {children}
-              </ReferenceContainer>
-            )}
-          </Reference>
+          <Reference>{({ ref }) => <ReferenceContainer innerRef={ref}>{children}</ReferenceContainer>}</Reference>
           {isHovering && (
             <Portal>
               <Popper placement={placement}>
@@ -77,5 +72,47 @@ class Tooltip extends React.Component<Props, State> {
     )
   }
 }
+
+const fadeIn = keyframes`
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+`
+
+export const ToolTipContainer = styled.div`
+  display: inline-block;
+`
+
+export const PopperContainer = styled.div`
+  padding: ${T('spacing.xs')}px;
+  font-size: ${T('font.size.sm')}px;
+  background-color: ${T('palette.white')};
+  box-shadow: 0 2px 8px ${T('palette.black16')};
+  animation: 0.3s ${fadeIn} ease-out;
+
+  &[data-placement|='top'] {
+    margin-bottom: 8px;
+  }
+
+  &[data-placement|='right'] {
+    margin-left: 8px;
+  }
+
+  &[data-placement|='bottom'] {
+    margin-top: 8px;
+  }
+
+  &[data-placement|='left'] {
+    margin-right: 8px;
+  }
+`
+
+export const ReferenceContainer = styled.div`
+  display: inline-block;
+  font-size: ${T('font.size.sm')}px;
+`
 
 export default Tooltip
